Resolve OS/Python Version Dependant Issues.
Two issues stem from an inequality in python versions:
 1. Due to PEP 3151, socket.error was made an alias of OSError in Python3.3 and greater. Those "pep" issues mentioned previously were PEP 3151 warnings, since checking "exc_type is socket.error" would no longer work because OSError changes it's instance's __class__ if instantiated with and errno and a string representation. To solve this issue in the server's handle_error, I simply used an issubclass check, since issubclass still evaluates to True when the classes are the same (in the case of Python2.7).
 2. Python 2 and 3 handle an empty raise statement differently when not directly within an except block. To keep them the same, six provides a convenient reraise function (the only downside being the reraise function call is added to the traceback)
And an issue dealing with underlying C library inequalities. On Windows, the `winerror` is approximately translated to the posix `errno` value, and that's fine. On linux, there would only be a posix errno value. However on OSX, errno values are different in the underlying C libraries, so assuming an arbitrary constant such as 32, 111, or 114 wouldn't match on OSX systems (or others that don't use these numbers to signal broken pipes, refused connections, and rereads). The underlying errno module is generated from C, so using constants from it are safe.