Fix and test do_evict and do_clear_cache
In Python2.7 and up, the classmethod decorator binds the defined function to the class that it is within, and stores the defined function in the __func__ attribute. Simply rebinding the already bound method to the RequestHandler does not let it interact with the RequestHandler's attributes. Instead, the inner __func__ has to be rebound via the classmethod decorator.

Also adds this to a test, which is renamed to handle both do_evict and the cache hit callback.